Adds Basic Post Route To Vote On A Location
Adds a basic route for users to cast votes for a particular location.
This route will rate a location based on awesomeness and busyness.
When there is no current rating in the database for a location, a new
Atmosphere rating is created.

TODO: Have Yelp service fetch information about the bar when creating
a new atmosphere rating.